// src/interfaces/user.ts
export interface objUser {
    id?: number;
    nombres?: string;
    apellidos?: string;
    fecha_cumple?: string;
    nro_documento?: string;
    tipo_documento?: string;
    email?: string;
    departamento?: string;
    provincia?: string;
    distrito?: string;
    movil?: string;
    tyc?: boolean;
    tyc_vacaciones?: boolean;
    recibirInformacion?: boolean;
}